Prevent occasional infinite loop when unwinding stack on ARM
If PC didn't change from the previous unwound frame and we made a guess
that SP is a frame pointer (last resort), we're now in an infinite loop
creating frames with SP increasing by cache->framesize, but in fact the
loop is unwinding a single stack frame over and over. Upon detecting
this condition use stack pointer from the current frame to create an
already existing id for this frame in order to force breaking the loop.